src/ui/server/controllers/insurance/eligibility/check-your-answers/index.ts
import { PAGES } from '../../../../content-strings';
import { TEMPLATES } from '../../../../constants';
import { INSURANCE_ROUTES } from '../../../../constants/routes/insurance';
import corePageVariables from '../../../../helpers/page-variables/core/insurance';
import getUserNameFromSession from '../../../../helpers/get-user-name-from-session';
import { Request, Response } from '../../../../../types';

const { ELIGIBLE_TO_APPLY_ONLINE } = INSURANCE_ROUTES.ELIGIBILITY;

export const PAGE_VARIABLES = {
  PAGE_CONTENT_STRINGS: PAGES.INSURANCE.ELIGIBILITY.CHECK_YOUR_ANSWERS,
};

export const TEMPLATE = TEMPLATES.INSURANCE.ELIGIBILITY.CHECK_YOUR_ANSWERS;

/**
 * get
 * Render the "Check your answers" page
 * @param {Express.Request} Express request
 * @param {Express.Response} Express response
 * @returns {Express.Response.render} Check your answers page
 */
export const get = (req: Request, res: Response) =>
  res.render(TEMPLATE, { ...corePageVariables({ ...PAGE_VARIABLES, BACK_LINK: req.headers.referer }), userName: getUserNameFromSession(req.session.user) });

/**
 * post
 * Redirect to the next part of the flow.
 * @param {Express.Request} Express request
 * @param {Express.Response} Express response
 * @returns {Express.Response.redirect} Next part of the flow
 */
export const post = (req: Request, res: Response) => res.redirect(ELIGIBLE_TO_APPLY_ONLINE);
